@@ -4,37 +4,44 @@
   "description": "ammobin.ca api",
   "main": "./dist/api/index.js",
   "scripts": {
-    "start": "node ./dist/api/index.js",
-    "start:worker": "node --tls-min-v1.0 ./dist/worker/index.js",
+    "clean": "rm -rf dist lambda",
+    "start": "node ./dist/api/api.js",
+    "start:worker": "node --tls-min-v1.0 ./dist/worker/redis-worker.js",
     "start:refresher": "node ./dist/refresher/index.js",
     "test:lint": "tslint ./src/**/*.ts",
     "test": "npm run test:lint && npm run build",
     "format": "pretty-quick",
     "build": "tsc",
-    "dev": "DEV=true ts-node-dev ./src/api/index.ts",
-    "dev:worker": "DEV=true ts-node-dev ./src/worker/index.ts",
-    "dev:refresher": "DEV=true ts-node-dev ./src/refresher/index.ts",
+    "build-lambda": "mkdir -p lambda/src && tsc --outDir lambda/src && cp package.json lambda && cp graphql.gql lambda && cd lambda && npm install --production && cd ..",
+    "dev": "DEV=true ts-node-dev ./src/api/api.ts",
+    "dev:worker": "DEV=true ts-node-dev ./src/worker/redis-worker.ts",
+    "dev:refresher": "DEV=true ts-node-dev ./src/refresher/redis-worker.ts",
     "generate-types": "gql2ts graphql.gql > src/graphql-types.ts && sh export-graphql-types.sh"
   },
   "types": "dist/index.d.ts",
   "author": "Simon Ramsay",
   "license": "MIT",
   "dependencies": {
-    "ammobin-classifier": "0.0.36",
-    "apollo-server": "^2.9.4",
+    "ammobin-classifier": "0.1.0",
+    "apollo-server": "^2.9.7",
     "apollo-server-cache-redis": "^1.1.4",
-    "apollo-server-hapi": "^2.9.4",
-    "apollo-server-plugin-response-cache": "^0.3.4",
+    "apollo-server-hapi": "^2.9.7",
+    "apollo-server-lambda": "^2.9.7",
+    "apollo-server-plugin-response-cache": "^0.3.5",
+    "aws-sdk": "^2.568.0",
     "axios": "^0.19.0",
     "boom": "^7.3.0",
     "cheerio": "^1.0.0-rc.3",
+    "chrome-aws-lambda": "^1.20.4",
     "delay": "^4.3.0",
     "fluent-logger": "^3.3.1",
     "graphql": "^14.5.8",
     "hapi": "^18.1.0",
+    "hapi-lambda": "^1.0.0",
     "jsonp": "^0.2.1",
     "moment": "^2.24.0",
     "node-cron": "^2.0.3",
+    "puppeteer-core": "^1.20.0",
     "redis": "^2.7.1",
     "rsmq-worker": "^0.5.2",
     "throat": "^5.0.0",
@@ -43,16 +50,17 @@
   "devDependencies": {
     "@types/boom": "^7.3.0",
     "@types/graphql": "^14.5.0",
-    "@types/hapi": "^18.0.2",
+    "@types/hapi": "^18.0.3",
     "@types/moment": "^2.13.0",
+    "@types/puppeteer": "^1.20.2",
     "gql2ts": "^1.10.1",
-    "husky": "^3.0.7",
-    "nodemon": "^1.19.2",
-    "prettier": "^1.18.2",
+    "husky": "^3.0.9",
+    "nodemon": "^1.19.4",
+    "prettier": "^1.19.1",
     "pretty-quick": "^1.11.1",
-    "ts-node-dev": "^1.0.0-pre.43",
-    "tslint": "^5.20.0",
-    "typescript": "^3.6.3"
+    "ts-node-dev": "^1.0.0-pre.44",
+    "tslint": "^5.20.1",
+    "typescript": "^3.7.2"
   },
   "husky": {
     "hooks": {
