@@ -1,25 +1,28 @@
-var winston = require('winston')
-var config = {
-  host: 'fluent',
-  port: 24224,
-  timeout: 3.0,
-  requireAckResponse: true, // Add this option to wait response from Fluentd certainly
-}
-var fluentTransport = require('fluent-logger').support.winstonTransport()
+import winston from 'winston'
 
-function createLogger(tag) {
+function createLogger(tag: string) {
+  const transports = []
+  if (process.env.FLUENT) {
+    let config = {
+      host: 'fluent',
+      port: 24224,
+      timeout: 3.0,
+      requireAckResponse: true, // Add this option to wait response from Fluentd certainly
+    }
+    let fluentTransport = require('fluent-logger').support.winstonTransport()
+    transports.push(new fluentTransport(tag, config))
+  }
+  if (process.env.DONT_LOG_CONSOLE !== 'true') {
+    transports.push(new winston.transports.Console())
+  }
+
+  if (transports.length === 0) {
+    console.warn('no longer transports configured....')
+  }
   return winston.createLogger({
-    transports: [
-      new fluentTransport(tag, config),
-      new winston.transports.Console(),
-    ],
+    transports,
   })
 }
 
-module.exports = {
-  apiLogger: createLogger('ammobin.api'),
-  workerLogger: createLogger('ammobin.api'),
-}
-
-export const apiLogger = createLogger('ammobin.api')
-export const workerLogger = createLogger('ammobin.api')
+export const apiLogger = createLogger('api')
+export const workerLogger = createLogger('api')
