@@ -12,10 +12,11 @@ import crypto from 'crypto'
 // used to encrypt request ip
 const secret = process.env.HASH_SECRET || Math.random().toString()
 
-import { typeDefs, resolvers } from './graphql'
+import { typeDefs, vendors, bestPrices } from './graphql'
 import { SOURCES, RELOAD_TYPES, AMMO_TYPES, TYPES } from '../constants'
 import { ItemType } from '../graphql-types'
-
+import { getItemsFlatListings, getScrapeResponses } from './shared'
+import { getRedisItems } from './redis-getter'
 const DEV = process.env.DEV === 'true'
 
 const client = redis.createClient({ host: 'redis' })
@@ -274,7 +275,16 @@ async function doWork() {
   try {
     const apolloServer = new ApolloServer({
       typeDefs,
-      resolvers,
+      resolvers: {
+        Query: {
+          vendors,
+          bestPrices,
+          itemsListings: (_, params) =>
+            getScrapeResponses(params, getRedisItems),
+          itemsFlatListings: (_, params) =>
+            getItemsFlatListings(params, getRedisItems),
+        },
+      },
       debug: DEV,
       tracing: DEV,
       cache: !DEV
