@@ -1,6 +1,10 @@
 import { SOURCES, AMMO_TYPES, RELOAD_TYPES, VENDORS } from '../constants'
 import * as helpers from '../helpers'
-import * as redis from 'redis'
+
+import {
+  IItemsFlatListingsOnQueryArguments,
+  IItemFlatListings,
+} from '../graphql-types'
 import {
   IItemsListingsOnQueryArguments,
   IItemListings,
@@ -9,63 +13,66 @@ import {
   IItemListing,
   SortOrder,
   SortField,
-  IBestPricesOnQueryArguments,
   IBestPrice,
 } from '../graphql-types'
 import { URL } from 'url'
-const client = redis.createClient({ host: 'redis' })
 
-export async function getBestPrices(
-  params: IBestPricesOnQueryArguments
-): Promise<IBestPrice[]> {
-  const type = params.type || ItemType.centerfire
-  const { calibres } = params
+export async function getBestPrices(): Promise<IBestPrice[]> {
+  return Promise.resolve(null)
+  // TODO: not used
+  // const type = params.type || ItemType.centerfire
+  // const { calibres } = params
 
-  const keys: string[] = SOURCES.map(s =>
-    calibres.map(subType => helpers.getKey(s, type, subType))
-  ).flat<string>(Infinity)
+  // const keys: string[] = SOURCES.map(s =>
+  //   calibres.map(subType => helpers.getKey(s, type, subType))
+  // ).flat<string>(Infinity)
 
-  const res: any = await new Promise((resolve, reject) =>
-    client.mget(keys, (err, res2) => (err ? reject(err) : resolve(res2)))
-  )
-  const results: IItemListing[][] = res.map(r => (r ? JSON.parse(r) : null))
-  const result = results
-    .reduce((final, result2) => {
-      return result2 && result2.length ? final.concat(result2) : final
-    }, [])
-    .reduce((response, item) => {
-      if (
-        !item ||
-        !item.subType ||
-        !item.unitCost ||
-        item.subType === 'UNKNOWN' ||
-        (calibres && !calibres.includes(item.subType))
-      ) {
-        return response
-      }
+  // const res: any = await new Promise((resolve, reject) =>
+  //   client.mget(keys, (err, res2) => (err ? reject(err) : resolve(res2)))
+  // )
+  // const results: IItemListing[][] = res.map(r => (r ? JSON.parse(r) : null))
+  // const result = results
+  //   .reduce((final, result2) => {
+  //     return result2 && result2.length ? final.concat(result2) : final
+  //   }, [])
+  //   .reduce((response, item) => {
+  //     if (
+  //       !item ||
+  //       !item.subType ||
+  //       !item.unitCost ||
+  //       item.subType === 'UNKNOWN' ||
+  //       (calibres && !calibres.includes(item.subType))
+  //     ) {
+  //       return response
+  //     }
 
-      if (!response[item.subType]) {
-        response[item.subType] = {
-          unitCost: Number.MAX_SAFE_INTEGER,
-          calibre: item.subType,
-          subType: item.subType,
-          type,
-        }
-      }
+  //     if (!response[item.subType]) {
+  //       response[item.subType] = {
+  //         unitCost: Number.MAX_SAFE_INTEGER,
+  //         calibre: item.subType,
+  //         subType: item.subType,
+  //         type,
+  //       }
+  //     }
 
-      response[item.subType].unitCost = Math.min(
-        response[item.subType].unitCost,
-        item.unitCost
-      )
+  //     response[item.subType].unitCost = Math.min(
+  //       response[item.subType].unitCost,
+  //       item.unitCost
+  //     )
 
-      return response
-    }, {})
+  //     return response
+  //   }, {})
 
-  return Object.values(result)
+  // return Object.values(result)
 }
 
 export async function getScrapeResponses(
-  params: IItemsListingsOnQueryArguments
+  params: IItemsListingsOnQueryArguments,
+  valueGetter: (
+    types: ItemType[],
+    subTypes: string[],
+    vendors: string[]
+  ) => Promise<IItemListing[]>
 ): Promise<IItemListings> {
   let {
     itemType,
@@ -98,19 +105,19 @@ export async function getScrapeResponses(
     pageSize = 100
   }
 
-  let types: ItemType[]
+  let itemTypes: ItemType[]
   switch (itemType) {
     case ItemType.reloading:
-      types = RELOAD_TYPES
+      itemTypes = RELOAD_TYPES
       break
     // default to ammo if nothing
     case ItemType.ammo:
     case undefined:
     case null:
-      types = AMMO_TYPES
+      itemTypes = AMMO_TYPES
       break
     default:
-      types = [itemType]
+      itemTypes = [itemType]
   }
 
   // figure out subset of keys to get (keys are stored by vendor)
@@ -124,25 +131,15 @@ export async function getScrapeResponses(
     )
   }
 
-  const keys: string[] = types
-    .map(t =>
-      vendors.map(s =>
-        (subType ? [subType] : helpers.itemTypeToStubTypes(t)).map(st =>
-          helpers.getKey(s, t, st)
-        )
+  const subTypes = subType
+    ? [subType]
+    : itemTypes.reduce(
+        (lst, t) => lst.concat(helpers.itemTypeToStubTypes(t)),
+        []
       )
-    )
-    .flat<string>(Infinity)
+  let results: IItemListing[] = await valueGetter(itemTypes, subTypes, vendors)
 
-  const results: IItemListing[][] = await new Promise((resolve, reject) =>
-    client.mget(keys, (err, rres: string[]) =>
-      err ? reject(err) : resolve(rres.map(r => (r ? JSON.parse(r) : null)))
-    )
-  )
-
-  // only filters out ammo without subType set (not setup for reloading yet)
   const result: IItemListing[] = results
-    .reduce((final, r) => (r ? final.concat(r) : final), [])
     .filter(
       r =>
         r &&
@@ -160,57 +157,53 @@ export async function getScrapeResponses(
         return 0
       }
     })
-
-  const itemsGrouped = result.reduce(
-    (r, item) => {
-      // if provided, filter out items that DONT match
-      if (
-        (subType && item.subType !== subType) ||
-        (query && !item.name.toLowerCase().includes(query.toLowerCase()))
-      ) {
-        return r
-      }
-
-      const key = item.subType + '_' + item.brand
-      if (!r[key]) {
-        r[key] = {
-          name: `${item.brand} ${item.subType}`,
-          subType: item.subType,
-          brand: item.brand,
-          minPrice: item.price,
-          maxPrice: item.price,
-          itemType: item.itemType,
-          minUnitCost: item.unitCost || 0,
-          maxUnitCost: item.unitCost || 0,
-          img: item.img,
-          vendors: [item],
-        } as IItemGroup
-      } else {
-        const val = r[key]
-        val.minPrice = Math.min(item.price, val.minPrice)
-        val.maxPrice = Math.max(item.price, val.maxPrice)
-
-        if (item.unitCost) {
-          if (val.minUnitCost === 0) {
-            val.minUnitCost = item.unitCost
-          }
-          val.minUnitCost = Math.min(item.unitCost, val.minUnitCost)
-          val.maxUnitCost = Math.max(item.unitCost, val.maxUnitCost)
-        }
-
-        if (val.img) {
-          // randomly pick group image
-          val.img = Math.random() > 0.5 ? item.img : val.img
-        } else {
-          val.img = item.img
-        }
-
-        val.vendors.push(item)
-      }
+  const itemsGrouped = result.reduce((r, item) => {
+    // if provided, filter out items that DONT match
+    if (
+      (subType && item.subType !== subType) ||
+      (query && !item.name.toLowerCase().includes(query.toLowerCase()))
+    ) {
       return r
-    },
-    <{ [key: string]: IItemGroup }>{}
-  )
+    }
+
+    const key = item.subType + '_' + item.brand
+    if (!r[key]) {
+      r[key] = {
+        name: `${item.brand} ${item.subType}`,
+        subType: item.subType,
+        brand: item.brand,
+        minPrice: item.price,
+        maxPrice: item.price,
+        itemType: item.itemType,
+        minUnitCost: item.unitCost || 0,
+        maxUnitCost: item.unitCost || 0,
+        img: item.img,
+        vendors: [item],
+      } as IItemGroup
+    } else {
+      const val = r[key]
+      val.minPrice = Math.min(item.price, val.minPrice)
+      val.maxPrice = Math.max(item.price, val.maxPrice)
+
+      if (item.unitCost) {
+        if (val.minUnitCost === 0) {
+          val.minUnitCost = item.unitCost
+        }
+        val.minUnitCost = Math.min(item.unitCost, val.minUnitCost)
+        val.maxUnitCost = Math.max(item.unitCost, val.maxUnitCost)
+      }
+
+      if (val.img) {
+        // randomly pick group image
+        val.img = Math.random() > 0.5 ? item.img : val.img
+      } else {
+        val.img = item.img
+      }
+
+      val.vendors.push(item)
+    }
+    return r
+  }, <{ [key: string]: IItemGroup }>{})
 
   // flatten map and sort groups by sortField + sortKey
   let res = Object.keys(itemsGrouped)
@@ -287,3 +280,119 @@ export async function getScrapeResponses(
     items,
   } as IItemListings
 }
+
+export async function getItemsFlatListings(
+  args: IItemsFlatListingsOnQueryArguments,
+  valueGetter: (
+    types: ItemType[],
+    subTypes: string[],
+    vendors: string[]
+  ) => Promise<IItemListing[]>
+): Promise<IItemFlatListings> {
+  let {
+    itemType,
+    page,
+    pageSize,
+    subType,
+    province,
+    vendor,
+    sortField,
+    sortOrder,
+    query,
+    brand,
+  } = args
+
+  if (!sortOrder) {
+    sortOrder = SortOrder.ASC
+  }
+
+  if (!sortField) {
+    sortField = SortField.minUnitCost
+  }
+
+  if (isNaN(page) || page < 1) {
+    page = 1
+  }
+
+  if (isNaN(pageSize) || pageSize < 1) {
+    pageSize = 25
+  } else if (pageSize > 100) {
+    pageSize = 100
+  }
+
+  let types: ItemType[]
+  switch (itemType) {
+    case ItemType.reloading:
+      types = RELOAD_TYPES
+      break
+    // default to ammo if nothing
+    case ItemType.ammo:
+    case undefined:
+    case null:
+      types = AMMO_TYPES
+      break
+    default:
+      types = [itemType]
+  }
+
+  // figure out subset of keys to get (keys are stored by vendor)
+  let vendors: string[] = VENDORS.map(v => v.name)
+  if (vendor) {
+    const vv = VENDORS.find(v => v.name === vendor)
+    vendors = vv ? [vv.name] : []
+  } else if (province) {
+    vendors = VENDORS.filter(v => v.provinces.includes(province)).map(
+      v => v.name
+    )
+  }
+  const subTypes = subType
+    ? [subType]
+    : types.reduce((lst, t) => lst.concat(helpers.itemTypeToStubTypes(t)), [])
+  let results: IItemListing[] = await valueGetter(types, subTypes, vendors)
+
+  // only filters out ammo without subType set (not setup for reloading yet)
+  const result: IItemListing[] = results.filter(
+    r =>
+      r &&
+      r.price > 0 &&
+      (!AMMO_TYPES.includes(itemType) ||
+        (r.subType && r.subType !== 'UNKNOWN')) &&
+      (!brand || (r.brand && r.brand.toLowerCase() === brand.toLowerCase())) &&
+      (!query || r.name.toLowerCase().includes(query.toLowerCase()))
+  )
+
+  // flatten map and sort groups by sortField + sortKey
+  let res = result.sort((a, b) => {
+    let aa = a[sortField]
+    let bb = b[sortField]
+    // put unknown unit costs at the bottom of the sort order
+    if (sortField === SortField.minUnitCost) {
+      if (aa <= 0) {
+        aa = Number.MAX_SAFE_INTEGER
+      }
+      if (bb <= 0) {
+        bb = Number.MAX_SAFE_INTEGER
+      }
+    }
+    const order = sortOrder === SortOrder.ASC ? 1 : -1
+    if (aa === bb) {
+      return 0
+    } else if (aa > bb) {
+      return 1 * order
+    } else {
+      return -1 * order
+    }
+  })
+  // paginate
+  // and then sort vendor listings within the selected groups
+  let items = res.slice((page - 1) * pageSize, page * pageSize)
+
+  return {
+    page,
+    pageSize,
+    pages: Math.ceil(res.length / pageSize),
+    items,
+  } as IItemFlatListings
+}
+
+// todo: merge the 2 functions together...
